Mostly working menus, maps load but do not render.
There are plenty of bugs, many probably due to the hacks
to get the right crtc offset.

Not working:
- Strip of wrong pixels down the side.
- Mouse control.
- Rendering maps.
- Help screen scrolling.

+++ id_ca.h
@@ -60,9 +60,9 @@
 
 typedef	struct
 {
-	long		planestart[3];
-	unsigned	planelength[3];
-	unsigned	width,height;
+	int32_t		planestart[3];
+	uint16_t	planelength[3];
+	uint16_t	width,height;
 	char		name[16];
 } maptype;
 
@@ -71,7 +71,7 @@ typedef	struct
 extern	byte 		*tinf;
 extern	int		mapon;
 
-extern	unsigned	*mapsegs[3];
+extern	uint16_t	*mapsegs[3];
 extern	maptype		*mapheaderseg[NUMMAPS];
 extern	byte		*audiosegs[NUMSNDCHUNKS];
 extern	void		*grsegs[NUMCHUNKS];
